Move type definitions to src/types
