+++ src/types/gdk.ts
