+++ src/types/gnome_desktop.ts
