+++ src/types/shell.ts
